Update patch set 9
Patch Set 9:

It does seem logical to adapt the https://github.com/openstack/ansible-role-container-registry/blob/master/tasks/docker-login.yml task for the overall login logic w/ docker or podman.  The ansible-role-container-registry is utilized in several other ci jobs already. 

One potential issue is that the login work that Gabriele is working on has to utilize zuul.secrets to login to the registry.  AFAICT this could be overcome by the tripleo-rdo-base/registry-login.yaml call the ansible-role-container-registry role w/ include task and passing the creds in the playbook.

wdyt?

Patch-set: 9
Label: Code-Review=0