Moving rdo jobs to ansible 2.8
This patch set the ansible-version for the rdo base jobs to ansible 2.8.
Notice this could not be done in the upstream base because it would move
also the upstream jobs to ansible 2.8, which is not the intention right
now.
This will also remove the need of use ansible_python_version on rhel 8
jobs, since ansible 2.8 has a fix to properly identify the python
interpreter in rhel 8.

Change-Id: Ib18917137bc5236551c03d0641c8582beb634f59