Proposing Martin Kopec as Tempest & tempestconf package core
Since Martin is taking care of tempest and python-tempestconf
RDO packaging by making sure each version of tempest and
tempestconf update does not break the CI and also leads to
good understanding of packaging.

It's time to give more power to him.

Thank you Martin for all the hard work.

Change-Id: Ice26ca7aa430ebbc1765ac3d8ad26bcffd8f962b
Signed-off-by: Chandan Kumar (raukadah) <chkumar246@gmail.com>