Update patch set 2
Patch Set 2:

I have a question. Imagine it runs in the weekend and some branch fails. Is running periodic on check enough to get back on track or we need to wait a whole week again to get results on the fix ? For example imagine the containers build fails, and all the jobs ther are skipped. How would we proceed ?

Patch-set: 2
Reviewer: Gerrit User 63 <63@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0