Update patch set 3
Patch Set 3:

> Good to hold atleast till rocky/queens are green and train pipeline
 > is ready, just few days.

+1 to wait for next promotion.

BTW, it's not too long to wait up to for one week after a blocker is fixed? by then, we may hit other week. Also between saturday and sunday few weeks will be merged. If moving to twice a week because of resources constraints, wouldn't be better wednesday and saturday or so?

Patch-set: 3