Report result to dlrnapi w/ actual used hash
In normal situations tripleo-ci-testing should be the hash name
to report results for a successful job.

This change enables the periodic jobs to report to dlrnapi w/ the
actual hash that has been used in the job.

The get-hash task is called again in the post playbook and it
overrides hash_info settings to call dlrnapi report result command.

Change-Id: Ic1e080d0074839f6c75095eba82d49dd92eb8bdb