Update patch set 7
Patch Set 7:

One detail to keep in mind: since this is a change to the config project, changes depending on this one will not be executed speculatively. You need to merge this (or move the definitions to the rdo-jobs project) to be able to test it.

Patch-set: 7
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0