Update patch set 1
Patch Set 1:

We are working with Joel to add a rdopkg reqcheck zuul job, but we need to add some changes to rdopkg. We would like to depends on those changes on rdopkg so we need to add rdopkg in the rdoproject.org zuul tenant.

Patch-set: 1
Reviewer: Gerrit User 112 <112@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0