Add periodic centos-7 jobs to build/push images
Currently fs002 is used to build/push images for centos 7.
fs002 deploys an underloud with a hwole lot of unnecessary
step to build the images. The RHEL-8 image build/push
jobs are faster and more targeted.

This review adds similar CentOS 7 job to replace fs002,
that can be used in the train and master pipelines.

Change-Id: Ie9472be8eba3eaa87e08a2cd56ac0e207be0f296