Update patch set 2
Patch Set 2: Code-Review-2

The main goal of the third party TripleO CI integration job was to trigger based on keywords.

We are only going to enable automatic job trigger only when podman and ceph-ansible developers and stackholders asks us to do it.

Reason: Both the developers does not know how to debug tripleo ci jobs and we need to wait for their confirmation before enabling the same.

We need to provide value to both the projects through weekly or weekend pipeline from our side once everything starts working fine then we can ask for the same to enable it.

Patch-set: 2