Assure we run podman-package-* jobs daily
Adds these jobs to the periodic pipeline, which runs them daily.
This should allow us to build a history of their reliability.

We keep the project names in order to be able to get a consolidated
view of their executions regardless what triggered them, like:

https://review.rdoproject.org/zuul/builds?job_name=podman-package-rhel-8

For stats we can easily filter by pipeline in order to remove noise
from check jobs.

Change-Id: Ie94570372ea4830053e750f4c985e4ae92dd02df