Update patch set 6
Patch Set 6: Workflow-1

> I like moving things to roles, but I'll need some help in
 > understanding why we have a role to promote consistent ->
 > tripleo-ci-testing and a role to promote components.

The two promotions have things in common for sure. However, some steps in the middle are totally different and exclusive, as for example check-criteria. Another difference is that the role has promote_name hardcoded, so these parameters need to be changeable. I'll study the role to see if we can reuse the same code for different types of promotions.

Patch-set: 6
Label: Workflow=-1