Use github-check keyword for triggering comment based rdo jobs
Currently all the jobs triggered through check-rdo runs in
both the github-{manual, check} pipelines as comment action is
called there.

We have a seperate pipeline to trigger the jobs based on comment
on PR. It will be easier to seperate and find out what jobs are
running based on comment and auto triggered.

For that github-check keyword is added for the same.

Change-Id: I724c4b107a61ff1ca2cc7ecb0eb4bb3d86519b0b
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>