Use check-github keyword for triggering jobs from podman prs
Currently check-rdo keyword is used on github pull request and their
jobs runs in both the github-{manual, check} pipelines.

We have a seperate pipeline `github-manual` to trigger the jobs
based on comment on PR. It will be easier to seperate and
find out what jobs are running based on comment and
auto triggered.

For that check-github keyword is added for the same.

https://tree.taiga.io/project/tripleo-ci-board/task/1433

Change-Id: I724c4b107a61ff1ca2cc7ecb0eb4bb3d86519b0b
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>