Add job to reqcheck upstream project
The purpose of this job is to run 'rdopkg reqcheck' from
upstream side. From distgit side, it's covered by the job
called 'rdopkg-reqcheck'.

Also, we add the possibility to rdopkg to make a review
with the appropriate changes. As rdopkg does not support
yet this kind of feature, we set the boolean 'make_review'
to false.

Change-Id: I68f332626490fbc63780b1c8d978e07bd0a43458