integration: promoted-components to tripleo-ci-testing
Add the job to promote/pin 'promoted-components' to
'tripleo-ci-testing' in the new integration pipeline.

Refactor the parent job to promote a hash be more generic,
so we can promote/pin any hash with the parent job and just
set the source/target hashes in the child job. Start w/
promoted-components-to-tripleo-ci-testing job, the entry
point for the new integration pipeline.

https://tree.taiga.io/project/tripleo-ci-board/task/1441

Change-Id: I647c5286dda0b70a9fcc727a7db03f44beb715c3