Make get-hash.sh work from new ci-config repo location
Originally merged in [1] and reverted in [2]. This adds the
required-project ci-config missing in [1].

In [3] the get-hash.sh script wass moved into the ci-config
to make it easier to work with. This wires up the invocations.
Tracked by tripleo-ci squad in [4]. Any job definition with
tripleo-ci-periodic-base/pre.yaml needs rdo-infra/ci-config
in required-projects so to call get-hash from the new location.

We can remove get-hash itself in a followup once we decide
it works correctly in the new location.

[1] https://review.rdoproject.org/r/24552
[2] https://review.rdoproject.org/r/24615
[3] https://review.rdoproject.org/r/24549
[4] https://tree.taiga.io/project/tripleo-ci-board/task/1442

Change-Id: I3763305ef6188bae0a5043bcd157a6855fe64d37