Make get-hash.sh work from new ci-config repo location
Originally merged in [1] and reverted in [2]. This adds the
required-project ci-config missing in [1].

In [3] the get-hash.sh script is moved into the ci-config repo
for easier debug and this wires up the invocations to use it.
Tracked by tripleo-ci squad in [4].

We can remove get-hash itself in a followup once we decide
it works correctly in the new location.

[1] https://review.rdoproject.org/r/24552
[2] https://review.rdoproject.org/r/24615
[3] https://tree.taiga.io/project/tripleo-ci-board/task/1442
[4] https://review.rdoproject.org/r/24549

Change-Id: I3763305ef6188bae0a5043bcd157a6855fe64d37