Fix criteria match from promote_source
Criteria file for component promotion has changed to
reflect the current design of common component-ci-testing
and promoted-component labels for all components.
Now the jobs are listed in a nested dict and promotion label
is not hardcoded anymore.

Splits check-promotion-criteria into separate files to be
more modular and testable in molecule without duplicate
blocks.

Includes fixes to molecule tests to check the new
criteria file with nested list of jobs for all components.

Change-Id: I3269e34d3b89d4c2be849efde25dde9a59e670c5