Add promote_source var to periodic base jobs for centos  8
The promote_source var is added in the relesae file but the pre
playbooks run before the release file is applied.
Therefore the promote_source var  needs to be defined at the
job level.

This review adds the var to the periodic base so jobs so the
chold jobs will all inherit that var.

Change-Id: I18f73f9efa01b413dfdd1b261173df9f18fc9c07