Update patch set 1
Patch Set 1:

Yatin, this is why the message was send. Anyone that relies on explicit centos-7 node is supposed to update his job to assure he is overriding the node.

Usually all jobs that rely on any specific nodeset, they have this system in their name and manually setting it.

There was never a guarantee that centos-7 will be the default nodeset forever, in fact the rdo 'contract' was that it offerts a current node by default.

The same kind of move happened upstream when they switched ubuntu version. If you do not mention it, it is going to change at some point.

Our deployment jobs are not going be affected by this because they use specific nodes already.

I would prefer to see a -1 when a specific job is identified as being affected by it.

Patch-set: 1
Reviewer: Gerrit User 371 <371@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0