Replace c7 full-tempest with api/scenario jobs
With this change we are replacing centos7 Standalone-full-tempest for
stable/train job in openstack-periodic-latest-released pipeline with
two new jobs:-

* periodic-tripleo-ci-centos-7-standalone-full-tempest-api-train
* periodic-tripleo-ci-centos-7-standalone-full-tempest-scenario-train

We made this change because Standalone-full-tempest sometimes goes
out in timeout state due to race conditions and concurrency issue.
So we have splitted it in two jobs to run faster and get better
results. Job defination is here[1]

We already successfully test run of above jobs for stable/train
here[2]

[1] https://review.rdoproject.org/r/#/c/26216/
[2] https://review.rdoproject.org/r/#/c/26217/

Change-Id: I3db9db0bfe01bf3180db627928116fb0387c8f1e