Port fedora rebuilds job to CentOS8
Currently these jobs are running on CentOS7 to rebuild
deps for both CentOS7 and CentOS8, but now we hit an
issue in [1] where new tags(like enhances/suggests etc) are
not recognized on CentOS7 as rpm package is old there(4.11) and
these tags are added in 4.14 which is there in CentOS8.
So good time to switch these jobs to CentOS8.

[1] https://review.rdoproject.org/r/#/c/26218/

Change-Id: Ib22108dae8f06c19b5ef584adb54bbaea4214c2a