Add image_upload var to post playbook
post playbook is used here to upload the overcloud & ipa images
to the image server. Since we have now seperate image building job
and fs02 job where image is built with in the job and consumed
there. uploading the both the images is not a good idea. So,
Adding image_upload var and set to false will help to disable
the image upload and helps to test the overcloud image with in
the job itself.

Change-Id: I5603577b82e5d80956dc42aab73013992bc8475f
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>