Update patch set 2
Patch Set 2:

Thank you for the review, Yatin.  I've made a patch to the spec to add python2 compatibility for Train, but I can't figure out how to add a train-rdo branch to the distgit repo.  I've tried changing the defaultbranch=train-rdo in .gitreview and then using the -R flag in git review, but my patch is being rejected.

Is there a process I need to complete to request/add the train-rdo branch to that repo?

Patch-set: 2