Update patch set 3
Patch Set 3: Code-Review+1

I'll leave others +3 it, especially if the keys need to be validated first.
Also, we've discussed this but for other reviewers: adding this role to the pre playbook means people may potentially tamper with the results of the CI. Even if the 3rd party CI is non-voting upstream, tampering it might hide problems down the line.

Patch-set: 3
Reviewer: Gerrit User 73 <73@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1