Tweak buildimage scripts - avoid convert error and re-use upload
This avoids the error in the convert-upload-undercloud by exiting
early if this is not a 'legacy' buildimage job (centos7 train and
earlier), in which case we don't need to convert and upload the
undercloud image.

Also makes it so we can re-use the upload-cloud-images.sh from the
downstream periodic rhel8/rhos17 jobs added with [1]. The job will
set the UPLOAD_URL so we guard here and don't override if set.

[1] https://tree.taiga.io/project/tripleo-ci-board/task/1728

Change-Id: Ic43d7b44792811268a09fea2ce29357231abd989