ovb-manage: set allocation pool for public_net
TripleO CI uses a static IP for the undercloud's public
interface. When using extra node in some job's there is
sometime a conflict, because the extra node get's the
static address allocated for the undercloud.

Use an allocation pool for the public_net to ensure
the extra node does not get a conflicting address.

The CI templates in tripleo-heat-templates uses these
allocation pools for the overcloud.
* [{"start": "10.0.0.10",
    "end": "10.0.0.50"}]
* [{"start": "2001:db8:fd00:1000::10",
    "end": "2001:db8:fd00:1000:ffff:ffff:ffff:fffe"}]

ovb-manage uses allocation pools that do not overlap.

Depends-On: https://review.opendev.org/733598
Change-Id: I16ad34cb61672db5322052a4baeaa2babe94ac9d
Closes-Bug: #1874418