Update patch set 6
Patch Set 6:

> I am not sure this will work out the box, upstream we'll probably
 > needs a way to ensure this CI can only block tripleo change,
 > through some sort of gerrit acl magic.

We can add the pipeline to whichever project we want in the https://review.rdoproject.org/r/#/c/28307/4/zuul.d/tripleo.yaml layout - we have it per project in there 

or did you mean something else?

I don't know either if it will work - at best it is just a permissions issue that we might resolve if the ci user can't -2. Alternatively we might try workflow -1? 
I think we can just try it my merging this and if it is blocking/breaking we will revert it.

Patch-set: 6