Use buildah login for new container build job
Containers are pushed with sudo buildah push
in new container build centos8 based job but
podman login is not working[1] so need to use
sudo buildah login.
Depends-On on [1] do not work as the playbook
is part of config project which is trusted so
need to get the patch merged to get it tested.

[1] https://review.opendev.org/#/c/739415
Change-Id: I2c916f7aded36711ba6118de34ae968990b53560