Remove admin user from ptl groups
We have detected an issue with the auth_pubtkt cookie size
(that bundle the list of user's groups) for the admin user.
Making the admin user unable to connect on SF.

This change removes the admin user from ptl groups. The acl
of those projects already includes rdo-superusers group where
the admin is part of.

Change-Id: Iaea2ee77fd07844da9851218a3c4414745bf18ad