Stop using deprecated port-security environment file
When running of openstack-virtual-baremetal master
this warning is in the logs:

WARNING: port-security environment file detected.
port-security is now the default. The existing
port-security environment files are deprecated and
may be removed in the future. Please use the
environment files without "port-security" in their
filename instead.

And since the environment manipulation is trying to
replace non-existing string? We end up trying to use
the default which is (../) One directory up.

Change-Id: Iec98994a918e17231bd97cba21df264d1592b595