Add jcapitao to RDO provenpackagers and superusers
In the last releases Joel has done a lot of good work in maintaining
RDO packages and dependencies and has demostrated his knowledge and
commitment on it, he deserves full rights on distgits.

This patch is adding Joel as provenpackager and superuser so that he has
+2/-2 votes and also be able to manually verify and merge reviews in
distgits if needed.

Congrats and thanks for your work, Joel!!

Change-Id: I7be33f3fcae81bde36dddceb3cd75b0e0fea44b4