Add support to set up routed networks OVB environment
Set the provision network to ctlplane

When deploying routed networks the ctlplane network for
the undercloud is always 'ctlplne'. Set the provision_net
name in the OVB stack to 'ctlplane' so that the build nodes
json command writes the "good" physical_network name in the
instackenv.json file.

Boolean option 'routed_networks' (default: false) added.
When enabled nodes_config 'zero_nodes' should be used,
'nodes_leaf1' and 'nodes_leaf2' defines the number of
virtual barematals in the respective leaf.

Change-Id: I173eb0c6f64c9c5a3a68c7e01338ef76638a406e