Rename fact role_args to role_extra_node_args
The role_args name is very generic. Let's make it more
specific so that it won't be confusing when additional
role useage for routed networks is added.

This is also a fact, so instead of setting a default
empty srting in defaults/main.yml, use |default('') in
tasks/generate_template.yml.

Change-Id: I6bbe993219cf9db3ec9ff39bc817b8e790fc7afb