Create repository for RDO website content
It has been decided in RDO meeting [1] to move the repository with RDO
website content under RDO gerrit instance.

This review is creating a new repository rdo-website for it. The group
of admins have been copied from current users with admin permissions in
the github project.

The replication to github is not been configured in this review as it
will the the last step when the content  of the repo is on sync with the
github one and CI is properly set.

[1] http://eavesdrop.openstack.org/meetings/rdo_meeting___2020_12_02/2020/rdo_meeting___2020_12_02.2020-12-02-14.01.txt

Change-Id: I303cca031e5397a75b4134f606738d19dad71adf