Update patch set 5
Patch Set 5:

> You can't :( It's a trusted repo.
 > What I'd suggest it's to extract this patch to be independent, then
 > to merge it so we'll could easily to revert it (and only it) if
 > need. After merge we'll run ovb jobs in CI and look for results
 > that they're ok. Possibly in a weekend. I can take a look in Sunday
 > that everything is ok.

So, I did some rebasing/dropping to change the order of the patches to move this and https://review.rdoproject.org/r/#/c/30490/7 to the end of the patch list.

If we can merge the patches before it, then do the last two over the weekend. I think we can also squash the last two into one patch to make the revert easy if we need that.

Patch-set: 5
Label: Code-Review=0