Add jobs to validate CBS builds for CentOS Stream 8
We add the jobs to validate and build the final package in
CBS for CentOS Stream 8. We also adapt playbook scripts

We use the ansible fact "ansible_distribution_release" from
the ansible upstream patch [1] as a job var. When we bump
our ansible version to the minimal one including this patch,
we'll be able to remove this job variable.

[1] https://github.com/ansible/ansible/pull/73034

Change-Id: Ia144fd5be99ae9a297587f0fa1faee60e9be3e8c