Update patch set 1
Patch Set 1:

The current sizing is based on the following estimations:

- Up to 80 OVB jobs
- 5 networks per job
- 5 ports per network

If we increase to 6 networks and 6 ports/network we'd need:

- 480 networks
- 2880 ports

Which is pretty close to the current quota, and could potentially be more if all jobs are OVB (max-servers is 110). In that case we should increase the quota to:

- 660 networks
- 3960 ports

Is that ok?

Patch-set: 1