Switch u-c update job to c8-stream
After recent gerrit upgrade the job which is running
on CentOS7 fails with:
git: 'interpret-trailers' is not a git command. See 'git --help

To resolve this git>=2.2 is needed which is not available in
CentOS7, let's switch this job to CentOS8-Stream where
it works fine.

Also update permissions for .ssh/config as in CentOS8-Stream
it reports Bad file permissions without it.

Change-Id: I2b0a5395781b1c02054d7f3ad3534cbf6866c7a2