Switch buildsys-tags-update and uc-deps-update jobs to c8-stream
After recent gerrit upgrade the job which is running
on CentOS7 fails with:
git: 'interpret-trailers' is not a git command. See 'git --help

To resolve this git>=2.2 is needed which is not available in
CentOS7, let's switch this job to CentOS8-Stream where
it works fine.

Also update permissions for .ssh/config as in CentOS8-Stream
it reports Bad file permissions without it and update
python3-PyYAML package name.

Change-Id: I3867eb80b5bc387f2073ee35c2241e64933e5683