Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 2: Code-Review+2
> 
> let's first stop the bleeding of the securitygroup resource and then rework the flow here.

As long as we're sure the extra node securitygroup is the one we are bleeding ... I'm worried it is the undercloud security group. But I guess you guys have ways to check that in the tenant.

Patch-set: 2
Reviewer: Gerrit User 469 <469@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1