Do not use -pinned DLRN jobs for tht-compat, remove CBS jobs
The tripleo-heat-templates-compat package is only build on Train
and older, so we do not need to use the -pinned DLRN jobs. Also,
make sure the CentOS 8 job is only executed on the train-rdo branch,
since rpm-master is not active anymore.

Remove CBS jobs, too, since we have not build the package in CBS
since Rocky, and there are no plans for new releases in old branches.

Change-Id: I6581dc9649cec14c3fddd00d5a161246e4289154