[WIP] Allow enabling baremetal-config drive for OVB BM's
Enabling config drive is generally a problem for OVB
BMs. But there is a use-case where OVB can be used to
set up BM's with a full cloud-image, and use the nodes
as pre-deployed server. In this case enabling config
drive is useful to get the initial network up on the
nodes.

Rename the parameter used to set config drive for the
extra node to 'extra_node_enable_config_drive' and default
it to:
  cloud_settings[cloud_name]['baremetal_config_drive']|default(false)

Change-Id: Iaaf0ec47f31e37233cef0cd88c63ea6174d297d4