[WIP] Allow enabling baremetal-config drive for OVB BM's
Enabling config drive is generally a problem for OVB
BMs. But there is a use-case where OVB can be used to
set up BM's with a full cloud-image, and use the nodes
as pre-deployed server. In this case enabling config
drive is useful to get the initial network up on the
nodes.

Change-Id: Iaaf0ec47f31e37233cef0cd88c63ea6174d297d4