Load jobs from upstream devstack project
As part of CS9 preparation, we should be able to run devstack jobs in
CS9 nodes in SF zuul. This change is loading the job definition from
devstack upstream project so that we can reuse them and override what
needed locally.

Note that some of the jobs in devstack have as parent tempest-full-py3
which is in tempest, so i guess i need to add both.

Change-Id: I9e051a063403a4f91b92c3b2b415842c5909e5fd