Added run.yaml playbook for compose promoter
This patch adds the run.yaml playbook to
execute the 'ssh-agent' and 'compose-promoter' role.

The same playbook is aslo defined in the ci-config repo [1]
but we can't import the playbook from another repo.
We also tried to run the job in ci-config but that
failed to fetch the secret because the secrets are defined in
config.

[1]: https://github.com/rdo-infra/ci-config/blob/master/playbooks/tripleo-ci-promote-centos-compose/run.yaml

Change-Id: I2150aae28dc898a3c16497e4a99dad04179cf558