[temp][cs9] Print the workspace name
In CS9, the zuul user name is zuul-worker and for cs8, it is
zuul user. When hash_info.sh is created and get sourced in
check for previous promotions task, workspace is pointing
to /home/zuul/workspace but it is not there leading
to failure.

Adding debug output to find out the issue.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I10813996524a757d0e162b20896f898bd1be6c73