[promote_hash]Set WORKSPACE shell var value
In CS9, ansible_user_dir is pointing to /home/zuul-worker
and cs8, it is pointing to /home/zuul.

When check-commit-hash-promote-target.sh script is called, it
looks for $WORKSPACE value which points to /home/zuul but
on cs9, it does not exists which leads to failure.

Explicitly setting the same fixes the issue.
It also removes debug info introduced in this patch
https://review.rdoproject.org/r/c/config/+/36193

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I2028e6357c744756acb9127553e9336e21adf7cd