Set WORKSPACE shell var value for check reported jobs
In CS9, ansible_user_dir is pointing to /home/zuul-worker
and cs8, it is pointing to /home/zuul.

When dlrnapi_check_reports.sh script is called, it
looks for $WORKSPACE value which points to /home/zuul but
on cs9, it does not exists which leads to failure.

Explicitly setting the same fixes the issue.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Id1f7f55cda680a1207be040040a5f8eeb68c4d3b