Remove removal of tmp files in cbs validation jobs
As part of CBS validation, we install packages in an alternative root
location /tmp/root. In long installation, it can take more that 15
minutes and the systemd tmp clean service remove content in the
directory leading to installation errors.

This patch removes the tmp files removal as done in the past for some
other tripleo jobs [1].

[1] https://review.rdoproject.org/r/c/config/+/28041

Change-Id: I48396c3404c27b64024f722f3a9aa63324eb8b3b