[opstools] Add ``opstools-distgit-cbs-base`` job
In order to make the opstools jobs easier to find, they're being moved to rdo-jobs.
This base job is being added because the cbs_opstools secret is needed,
and can only be accessed from jobs defined in this repo.
Since the secret was only used in the pre-run stage, it can
be added into a job on its own, and jobs in the rdo-jobs
repo can inherit from it.

Change-Id: Iba0edfe7d99e409103b1e507928da5f1134da3fa