Remove CBS jobs from train and ussuri
Train and Ussuri and in Extended Maintenance and not creating more tags
and the releases in CloudSIG not longer supported so let's remove the
jobs as they can cause CI issues in certain cases. Note that yoga is the
last RDO release supporting CS8 so we don't need to have generic regex
for future releases.

Change-Id: I66b8dad55889b97dc6380dabe686a2031dbee037