Renaming the registry secrets and the build container quay jobs
Renamed the registry secrets to make them more meaningful.
Added -secrets to the c9 and c8 build container quay jobs. Since we use different credentials for the different registries in quay, each branch needs a different secret.
So we will include all the secrets that branches with centos 9 need in the -centos-9-secrets job (same goes for c7, c8 and other versions)
The branch specific jobs will be defined in rdo-jobs.

Change-Id: Ia6088fda2bba034218576c05031128f7e1862309