Adds ceph daemon promotion jobs and playbooks
This patch adds a new promotion job for ceph-daemon, which
copies ceph-daemon container image to a list of destination registries.
Both source and destination registries are hardcoded to avoid unwanted
content in our destination registry.
The job will first copy the image to the destination registry and after
that, with the help of artifact_promoter, will create a new remote
artifact in rdo server to point to the latest promoted tag.
This first patch is using 'quay.rdoproject.org' as destination registry.

Change-Id: Ibda2a2171e8af5f234975011e7ee1408dbac7f27
Signed-off-by: Douglas Viroel <viroel@gmail.com>