Report CBS build task url to gerrit comments
Very often, when working with CBS tasks, we need to look for the taskid
in job logs. This is tedious, so I'm proposing in this patch to report a
comment to gerrit reviews with the URL of the CBS build task so that we
can find it directly in the review web instead of going to the logs.

Depends-On: https://review.rdoproject.org/r/c/gating_scripts/+/44243
Change-Id: I75a1b53025456dfa6e2ec3c9180b5ea371831c22