Provide custom DNS and disable systemd-resolved service
Note, It is taken from this review
https://softwarefactory-project.io/r/c/software-factory/sf-operator/+/25926
```
The systemd-resolved service is changing the nameservers in
/etc/resolv.conf file from DNS that are set in Openstack subnet
to 127.0.0.53 which is using later cloud provider DNS, that
sometimes does not work as expected.
```

Jobs running on vexxhost goes into RETRY_LIMIT while
do dnf cache update. It happened due to change in DNS.
More info is in this bug:
https://bugs.launchpad.net/tripleo/+bug/1983817

It will help to fix the issue.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ie6f7d8e16addfcf3d068cb297ac784e758b42688