Do not ignore errors in roles/ovb-manage/tasks/ovb-delete-stack
Using ignore_errors results in a nasty failed task which is misleading
when ignore_errors is set to true. Instead, we should be clearer about
the intent and never fail the tasks.

The follow on tasks that use the result always focus on the rc/output
so there does not seem to be any negative side-effect to this change.

Change-Id: I3b4d449f877eebec94210e703b3a2cd2e7e9b8f6