Replace constraint_file link - connection timeouts
Master and Wallaby jobs are hitting retries while
trying to access the upper-constraints.txt file.

This patch changes the link where that file is
accessed to prevent the network access timeout.

We have changed the default constraint_file var to
use stable/train upper-constraints.txt to avoid
breakage on cs8.

Change-Id: I3417dd12cdfda64588b6b3fbf4a50558fc395e5c
Closes-Bug: #1990556