Add task to generate hash_info with override for distro
For the 8-9 mixed OS job running in the c8/wallaby line, the
node on which hash_info is generated and delorean reports from
is c9. This causes the related-bug where we report for c9 hash
instead of the c8 hash. This adds an override for distro so we
fetch the right hash.

Change-Id: I90e399c5e513bcacc4dfed1df1f1149d076c48e7
Related-Bug: https://bugs.launchpad.net/tripleo/+bug/1990012