Use upper-constraints.txt from clonned openstack/requirements
https://review.rdoproject.org/r/c/config/+/45232
Replaced constraint_file link to use it from opendev.org
to fix connection timeouts.

But we are still seeing the same issue.

In order to fix it:

- https://review.rdoproject.org/r/c/rdo-jobs/+/45296 adds
  the  openstack/requirements under ovb required-projects jobs.

- This patch uses the constriant files from clonned
  openstack/requirements projects from the CI job workspace.
  It will avoid connection timeout issues.

While testing here,
https://review.rdoproject.org/r/c/testproject/+/45294/2#message-a230c36a8479bf81956c6d9d862c4de147be6174
All the releases are working fine but
cs8 wallaby jobs requires python-glanceclient as a requirements.

So added python-glanceclient as a requirements.

Closes-Bug: lp#1990556

Depends-On: https://review.rdoproject.org/r/c/rdo-jobs/+/45296

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ifdd4c7853de03d241909d6b802ac6a57f76e6b74