Simplify mixed os centos8 reporting fix and tweak for components
The block in pre.yaml was added for [1] to fix delorean reporting.
However after [2] and the addition of the post.yaml fix [3] the pre
is only needed/used by centos 8 *component* job. The depends-on
removes the mixed_os_stable_version variable from the c9 integration
and component job definitions. So here we can simplify and keep only
the centos8 case.

In the post also tweak the conditional for components so we don't
re-report in that case (currently no common c9/c8 component line).

[1] https://bugs.launchpad.net/tripleo/+bug/1990012
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/46119
[3] https://review.rdoproject.org/r/c/config/+/46130

Depends-On: https://review.rdoproject.org/r/c/rdo-jobs/+/46160
Change-Id: I77adee29384ce3e580eedbbcfacafea12152eeb4