Simplify the mixed os centos8 reporting fix - keep only 8 case
The block was added as of the fix for [1] to ensure correct delorean
reporting. However after [2] and the addition of [3] this fix is now
only needed/used by the centos 8 *component* job. The depends-on
removes the mixed_os_stable_version variable from the c9 integration
and component job definitions. So here we can simplify and keep only
the centos8 case.

[1] https://bugs.launchpad.net/tripleo/+bug/1990012
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/46119

Depends-On: https://review.rdoproject.org/r/c/rdo-jobs/+/46160
Change-Id: I77adee29384ce3e580eedbbcfacafea12152eeb4