Remove uc file of previous iteration
We need to remove the upper-constraints file of previous
iteration otherwise if the next curl command fails for whatever
reason, the "update_uc" command will run with previous uc file,
leading to wrong update.

Plus, we add some retry mechanism to the curl command to maximize
chance of getting the upper-constraints file.

Change-Id: Ib4ca6e3c897bafd00078d187b69a2f8f8f6dc6bb