Run crc playbook on specific host
In a multinode job, we have different nodes. For example[1]
It is not necessary all the nodes using crc label.
If we run the crc playbook using all hosts, then it is going to fail
on other nodes not having crc label.

It fixes the playbook to run on controller node only.

[1]. https://review.rdoproject.org/r/c/rdo-jobs/+/46369

Signed-off-by: Chandan Kumar <chkumar@redhat.com>
Change-Id: I8a9a1ec4cde9cb6653073613bb086fe93fd29f7b