Update generation of changeid for the trigger job changes
As noticed in the trigger change at [1] we are causing jobs to
abort as we are updating the same trigger change before zuul can
run and report on the earlier change.

This started after the commit at [2] which intended to re-use
the trigger changes instead of creating a new one each time.

This change adds the branch, os version and component into the
change ID generation. This should result in a single test change
per branch/os pair and per component to prevent the jobs being
aborted before they run.

[1] https://review.rdoproject.org/r/c/tripleo-downstream-trigger-nested-virt/+/46510/
[2] https://github.com/rdo-infra/review.rdoproject.org-config/commit/4d9268cff8f9a31a4a7a53119f62f4206b5f21be

Change-Id: I3792c353f99250dbb9e66d6a32cf195fb38fa3ec