DNM Provide softwarefactory-project.io connection
The Microshift role is available in software-factory/sf-infra
project. Until the Microshift role is not merged in upstream [1],
the TripleO team would like to verify Openstack k8s project not
on CRC but also on the Microshift service.
The base-microshift job has been ported [2], but the job requires
to pull sf-infra project.

[1] https://review.opendev.org/c/zuul/zuul-jobs/+/876081
[2] https://review.rdoproject.org/r/c/config/+/47249

Change-Id: I1448fb44506bd66724658ff2e3744612a12e4507