Prepull images to CRC before removing the pull-secret
Till now a whole OpenStack operator deployment was needed to
gather all the private hosted container images before removing
the pull secret from the zuul VM in pre-run. This commit aims
to copy the neccesary pull-secret to the CRC VM and disable
any kind of access to the CRC since that secret copy. As the
CRC is using crio with selinux enabled is not possible to
access the VM FS and thus the secret from a pod neither.

Change-Id: Ief967fe7a53eb6379ac47a5beb1cc22ec881b7d7