Add localhost to resolv.conf for dataplane job
When running a data-plane adoption job with a multinode configuration
and network isolation, the /etc/resolv.conf does not contain an entry
for localhost, which interferes with dns resolution for the crc cluster
and prevents pod creation. This change checks the file and if the
localhost entry is missing, adds it.

Change-Id: Ibcde5d38e9dcd622f1bd1c223eeb27f49b9ee058