Call crc_insecure_registry playbook inside simple start
In order to start using base-simple-crc as a parent in EDPM
job which depends on content-provider, we need to configure
the inscure registry with in crc vm.

By running crc_insecure_registry playbook within simple-start
playbook configures the insecure registry.

Note: it also drops duplicate task for crc login task.
as it is already called in start and simple start playbook.

Signed-off-by: Chandan Kumar <chkumar@redhat.com>
Change-Id: I7c19b4f8f2dac46b239dff5a0b0a38d0480a0285